Created portfolio folder in storage, "add portfolio" button now saves new portfolio files as markdown
