--- resources/themes/ghost/templates/portfolio.php
@@ -1,158 +1,0 @@
-<?php
-namespace Ziki\Core;
-
-use Parsedown;
-use Mni\FrontYAML\Parser;
-use KzykHys\FrontMatter\FrontMatter;
-use Symfony\Component\Finder\Finder;
-use KzykHys\FrontMatter\Document as Doc;
-
-/**
- *	The document class holds all properties and methods of a single page document.
- *
- */
-
-class Portfolio
-{
-
-    //define an instance of the symfony clss
-    //define an instance of the frontMatter class
-
-    protected $file;
-
-    public function __construct($file)
-    {
-        $this->file       = $file;
-    }
-
-    public function file()
-    {
-        return $this->file;
-    }
-
-    //for creating markdown files
-    public function createportfolio($title, $content, $image)
-    {
-        $time = date(DATE_RSS, time());
-        $unix = strtotime($time);
-        // Write md file
-        $document = FrontMatter::parse($content);
-        $md = new Parser();
-        $markdown = $md->parse($document);
-
-        $yaml = $markdown->getYAML();
-        $html = $markdown->getContent();
-        //$doc = FileSystem::write($this->file, $yaml . "\n" . $html);
-
-        $yamlfile = new Doc();
-        $yamlfile['title'] = $title;
-        if ($tags != "") {
-            $tag = explode(",", $tags);
-            $put = [];
-            foreach ($tag as $value) {
-                array_push($put, $value);
-            }
-            $yamlfile['tags'] = $put;
-        }
-        if (!empty($image)) {
-            foreach ($image as $key => $value) {
-                $decoded = base64_decode($image[$key]);
-                $url = "./storage/images/portfolio/" . $key;
-                FileSystem::write($url, $decoded);
-            }
-        }
-
-        $yamlfile['post_dir'] = SITE_URL . "/storage/portfolio/{$unix}";
-        $yamlfile['post_dir'] = SITE_URL . "/storage/portfolio/{$unix}";
-        $yamlfile['image'] = "./storage/images/portfolio/" . $key;
-
-
-        // create slug by first removing spaces
-        $striped = str_replace(' ', '-', $title);
-        // then removing encoded html chars
-        $striped = preg_replace("/(&#[0-9]+;)/", "", $striped);
-        $yamlfile['slug'] = $striped . "-{$unix}";
-        $yamlfile['timestamp'] = $time;
-        $yamlfile->setContent($content);
-        $yaml = FrontMatter::dump($yamlfile);
-        $file = $this->file;
-        $dir = $file . $unix . ".md";
-        //return $dir; die();
-        $doc = FileSystem::write($dir, $yaml);
-        // if (!$extra) {
-        //     if ($doc) {
-        //         $result = array("error" => false, "message" => "Portfolio created successfully");
-        //         $this->createRSS();
-        //     } else {
-        //         $result = array("error" => true, "message" => "Fail while creating, please try again");
-        //     }
-        // } else {
-        //     if ($doc) {
-        //         $result = array("error" => false, "message" => "Draft saved successfully");
-        //     } else {
-        //         $result = array("error" => true, "message" => "Fail while publishing, please try again");
-        //     }
-        // }
-
-        return $result;
-    }
-
-    //get post
-    public function getportfolio()
-    {
-        $finder = new Finder();
-
-        // find all files in the current directory
-        $finder->files()->in($this->file);
-        $posts = [];
-        if ($finder->hasResults()) {
-            foreach ($finder as $file) {
-                $document = $file->getContents();
-                $parser = new Parser();
-                $document = $parser->parse($document);
-                $yaml = $document->getYAML();
-                $body = $document->getContent();
-                //$document = FileSystem::read($this->file);
-                $parsedown  = new Parsedown();
-                $tags = isset($yaml['tags']) ? $yaml['tags'] : '';
-                $title = isset($yaml['title']) ? $parsedown->text($yaml['title']) : '';
-                $slug = $parsedown->text($yaml['slug']);
-                $image = isset($yaml['image']) ? $parsedown->text($yaml['image']) : '';
-                $slug = preg_replace("/<[^>]+>/", '', $slug);
-                $image = preg_replace("/<[^>]+>/", '', $image);
-                $bd = $parsedown->text($body);
-                ////
-                preg_match('/<img[^>]+src="((\/|\w|-)+\.[a-z]+)"[^>]*\>/i', $bd, $matches);
-                $first_img = false;
-                if (isset($matches[1])) {
-                    // there are images
-                    $first_img = $matches[1];
-                    // strip all images from the text
-                    $bd = preg_replace("/<img[^>]+\>/i", " (image) ", $bd);
-                }
-                $time = $parsedown->text($yaml['timestamp']);
-                $url = $parsedown->text($yaml['post_dir']);
-                $content['title'] = $title;
-                $content['body'] = $this->trim_words($bd, 200);
-                $content['url'] = $url;
-                $content['timestamp'] = $time;
-                $content['tags'] = $tags;
-                $content['slug'] = $slug;
-                $content['preview_img'] = $first_img;
-                //content['slug'] = $slug;
-                $file = explode("-", $slug);
-                $filename = $file[count($file) - 1];
-                $content['filename'] = $filename;
-                //content['timestamp'] = $time;
-                $content['image'] = $image;
-                $content['date'] = date('d M Y ', $filename);
-
-                array_push($posts, $content);
-            }
-            krsort($posts);
-            return $posts;
-        } else {
-            return false;
-        }
-    }
-}
